+++ CHANGES.md
@@ -3,6 +3,8 @@
 It is now possible to mark fields (including operations) and enum values
 as deprecated.
 
+Compiled schemas now print and pretty-print as `#CompiledSchema<>`.
+
 ## 0.24.0 -- 30 Jan 2018
 
 Added the FieldResolver protocol that allows a Clojure record, such as a